# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trip API
  version: 0.1.0
  summary: Backend API for the Trip iOS application.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [Auth]
      responses:
        "201":
          description: User registered.
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Auth]
      responses:
        "200":
          description: User logged in.
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      tags: [Auth]
      responses:
        "200":
          description: Session refreshed.
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/auth/yandex:
    post:
      operationId: loginWithYandex
      tags: [Auth]
      responses:
        "200":
          description: Yandex OAuth token exchanged for backend tokens.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Auth]
      responses:
        "200":
          description: Current user.
        "401":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateMe
      tags: [Auth]
      responses:
        "200":
          description: User updated.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VERSION_CONFLICT
            - INVALID_CREDENTIALS
            - TOKEN_EXPIRED
            - TRIP_DATE_CONFLICT
            - MEMBER_ALREADY_EXISTS
            - EXPENSE_SHARES_MISMATCH
            - INTERNAL_ERROR
            - NOT_IMPLEMENTED
        message:
          type: string
        request_id:
          type: string
        details:
          type: array
          items:
            type: object
            required: [field, reason]
            properties:
              field:
                type: string
              reason:
                type: string